Migrate getClassroomIdByTopicId function
